+++ src/chart/options/options.h
@@ -28,14 +28,38 @@ namespace Vizzu::Gen
 
 class Options
 {
+	using ChannelIdType = std::underlying_type_t<ChannelId>;
+
 public:
+	enum class LegendId : ChannelIdType {
+		color = static_cast<ChannelIdType>(ChannelId::color),
+		lightness = static_cast<ChannelIdType>(ChannelId::lightness),
+		size = static_cast<ChannelIdType>(ChannelId::size)
+	};
+
+	static_assert(Refl::enum_names<LegendId>.size() == 3);
+	static_assert(std::ranges::all_of(Refl::enum_names<LegendId>,
+	    [](std::string_view name)
+	    {
+		    return static_cast<ChannelIdType>(
+		               Refl::get_enum<LegendId>(name))
+		        == static_cast<ChannelIdType>(
+		            Refl::get_enum<ChannelId>(name));
+	    }));
+
 	using MarkerId = uint64_t;
 	using Heading = ::Anim::Interpolated<std::optional<std::string>>;
-	using LegendType = Base::AutoParam<ChannelId>;
+	using LegendType = Base::AutoParam<LegendId>;
 	using Legend = ::Anim::Interpolated<LegendType>;
 	using OrientationType = Base::AutoParam<Gen::Orientation>;
 	using Orientation = ::Anim::Interpolated<OrientationType>;
 	using MarkersInfoMap = std::map<uint64_t, MarkerId>;
+
+	friend bool operator==(const Options::LegendId &,
+	    const ChannelId &);
+
+	[[nodiscard]] static ChannelId toChannel(
+	    const Options::LegendId &);
 
 	Options() = default;
 
@@ -153,7 +177,7 @@ private:
 	Channels channels;
 
 	[[nodiscard]] Gen::Orientation getAutoOrientation() const;
-	[[nodiscard]] std::optional<ChannelId> getAutoLegend() const;
+	[[nodiscard]] std::optional<LegendId> getAutoLegend() const;
 	static void setMeasureRange(Channel &channel, bool positive);
 	static void setRange(Channel &channel,
 	    ChannelExtrema min,